+++ lib/dst.dart
@@ -1,8 +1,10 @@
+import 'package:dst/dst_transition.dart';
+
 import 'dst_platform_interface.dart';
 
 class Dst {
 
-  Future<DateTime?> nextDaylightSavingTransitionAfterDate(DateTime date, String timeZoneName) {
+  Future<DstTransition?> nextDaylightSavingTransitionAfterDate(DateTime date, String timeZoneName) {
     return DstPlatform.instance.nextDaylightSavingTransitionAfterDate(date, timeZoneName);
   }
 }